feat: ✨ Create verify_platform function to get/create Platform to ensure standardized.
This should ensure that all Platforms have the name, slug, and NAPALM driver filled in using the netutils mappers. The standard is to have the Platfrom name be the Ansible collection, slug is netmiko driver, and NAPALM driver is obvious. If none of these apply it uses the name passed to it.